Pull common code out of D8S transformer into a common transformer
also create a prayer transformer for both D8 and D8S, currently
D8 will not use the new transformer until master has been merged
up into this branch to get latest changes and avoid conflicts.

Edit D8SFormToCaseTransformer to use new common and prayer
transformers.
