+++ src/main/java/uk/gov/hmcts/divorce/bulkscan/transformation/CommonTransformer.java
@@ -1,0 +1,117 @@
+package uk.gov.hmcts.divorce.bulkscan.transformation;
+
+import com.fasterxml.jackson.core.type.TypeReference;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang3.StringUtils;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
+import uk.gov.hmcts.divorce.bulkscan.validation.data.OcrDataFields;
+import uk.gov.hmcts.divorce.divorcecase.model.ApplicationType;
+import uk.gov.hmcts.divorce.divorcecase.model.CaseData;
+import uk.gov.hmcts.divorce.endpoint.data.OcrValidationResponse;
+
+import java.util.List;
+import java.util.Map;
+
+import static org.apache.commons.collections4.ListUtils.union;
+import static org.apache.commons.lang3.BooleanUtils.toBoolean;
+import static org.apache.commons.lang3.StringUtils.isNotEmpty;
+import static org.springframework.util.CollectionUtils.isEmpty;
+import static uk.gov.hmcts.ccd.sdk.type.YesOrNo.NO;
+import static uk.gov.hmcts.divorce.divorcecase.model.ApplicationType.JOINT_APPLICATION;
+import static uk.gov.hmcts.divorce.divorcecase.model.ApplicationType.SOLE_APPLICATION;
+
+@Component
+@Slf4j
+public class CommonTransformer {
+
+    @Autowired
+    private ObjectMapper mapper;
+
+    public static final String OCR_FIELD_VALUE_BOTH = "both";
+    public static final String TRANSFORMATION_AND_OCR_WARNINGS = "transformationAndOcrWarnings";
+
+    public List<String> verifyFields(TransformationDetails transformationDetails, List<String> transformationWarnings) {
+
+        CaseData caseData = transformationDetails.getCaseData();
+        OcrDataFields ocrDataFields = transformationDetails.getOcrDataFields();
+
+        verifyRespondentEmailAccess(ocrDataFields, transformationWarnings);
+        verifyServeOutOfUK(caseData.getApplicationType(), ocrDataFields, transformationWarnings);
+        verifyHowApplicationIsServed(caseData.getApplicationType(), ocrDataFields, transformationWarnings);
+
+        return transformationWarnings;
+    }
+
+    public CaseData setLabelContentAndDefaultValues(CaseData caseData) {
+
+        caseData.getLabelContent().setApplicationType(caseData.getApplicationType());
+        caseData.getLabelContent().setUnionType(caseData.getDivorceOrDissolution());
+
+        caseData.getConditionalOrder().getConditionalOrderApplicant1Questions().setIsSubmitted(NO);
+        caseData.getConditionalOrder().getConditionalOrderApplicant1Questions().setIsDrafted(NO);
+        if (!caseData.getApplicationType().isSole()) {
+            caseData.getConditionalOrder().getConditionalOrderApplicant2Questions().setIsSubmitted(NO);
+            caseData.getConditionalOrder().getConditionalOrderApplicant2Questions().setIsDrafted(NO);
+        }
+
+        return caseData;
+    }
+
+    public ApplicationType getApplicationType(OcrDataFields ocrDataFields, List<String> warnings) {
+        boolean isSole = toBoolean(ocrDataFields.getSoleApplication());
+        boolean isJoint = toBoolean(ocrDataFields.getJointApplication());
+        if (isJoint && !isSole) {
+            return JOINT_APPLICATION;
+        } else if (isSole && !isJoint) {
+            return SOLE_APPLICATION;
+        } else {
+            warnings.add("Please review application type in the scanned form");
+            return SOLE_APPLICATION;
+        }
+    }
+
+    public Map<String, Object> transformCaseData(CaseData caseData, List<String> transformationWarnings,
+                                                 OcrValidationResponse ocrValidationResponse) {
+
+        Map<String, Object> transformedCaseData = mapper.convertValue(caseData, new TypeReference<>() {
+        });
+
+        List<String> combinedWarnings = isEmpty(ocrValidationResponse.getWarnings())
+            ? transformationWarnings
+            : union(ocrValidationResponse.getWarnings(), transformationWarnings);
+
+        transformedCaseData.put(TRANSFORMATION_AND_OCR_WARNINGS, combinedWarnings);
+
+        log.info("Transformed case data map {} ", transformedCaseData);
+
+        return transformedCaseData;
+    }
+
+    private void verifyHowApplicationIsServed(ApplicationType applicationType, OcrDataFields ocrDataFields, List<String> warnings) {
+        if (SOLE_APPLICATION.equals(applicationType)) {
+            if (StringUtils.isEmpty(ocrDataFields.getRespondentServePostOnly())
+                || StringUtils.isEmpty(ocrDataFields.getApplicantWillServeApplication())) {
+                warnings.add("Please review respondent by post and applicant will serve application in the scanned form");
+            }
+            if (StringUtils.isEmpty(ocrDataFields.getRespondentDifferentServiceAddress())) {
+                warnings.add("Please review respondent address different to service address in the scanned form");
+            }
+        }
+    }
+
+    private void verifyServeOutOfUK(ApplicationType applicationType, OcrDataFields ocrDataFields, List<String> warnings) {
+        if (SOLE_APPLICATION.equals(applicationType)
+            && (OCR_FIELD_VALUE_BOTH.equalsIgnoreCase(ocrDataFields.getServeOutOfUK())
+            || StringUtils.isEmpty(ocrDataFields.getServeOutOfUK()))) {
+            warnings.add("Please review serve out of UK in the scanned form");
+        }
+    }
+
+    private void verifyRespondentEmailAccess(OcrDataFields ocrDataFields, List<String> warnings) {
+        if (isNotEmpty(ocrDataFields.getRespondentOrApplicant2Email()) && StringUtils.isEmpty(ocrDataFields.getRespondentEmailAccess())) {
+            warnings.add("Please verify respondent email access in scanned form");
+        }
+    }
+}
